chore: moved some stuff fromm auth module to user
move /me and /update_username to user module and add add search and getter endpiont
